requirements.txt
google-generativeai
pandas
pytest
python-dotenv
streamlit
